added a new enemymanager + yolov8
