+++ .gitignore
@@ -1,3 +1,3 @@
-/yolov5/
+/other/yolov5/
 /venv/
