+++ requirements.txt
@@ -1,6 +1,6 @@
 numpy==1.26.3
 PyScreeze==0.1.29
-Pillow==10.2.0
+pillow>=10.3.0
 PyAutoGUI==0.9.54
 pytesseract==0.3.10
 sounddevice==0.4.6
@@ -9,7 +9,7 @@ keyboard==0.13.5
 mouse==0.7.1
 pandas==2.1.4
 opencv-python>=4.8.1.78
-requests==2.31.0
+requests>=2.32.0
 PyYAML==6.0.1
 ultralytics==8.1.1
 matplotlib==3.8.2
@@ -17,7 +17,7 @@ seaborn==0.13.1
 scipy==1.11.4
 setuptools
 pywin32
-tqdm==4.66.1
+tqdm>=4.66.3
 psutil==5.9.7
 thop==0.1.1-2209072238
 keras_ocr==0.9.3
